# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:5050",
  "servers": [
    {
      "url": "",
      "description": ""
    }
  ],
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/v1/user/": {
      "get": {
        "tags": [
          "User Control"
        ],
        "summary": "Get list of all users in database",
        "description": "",
        "parameters": [
          {
            "name": "phoneNumber",
            "in": "query",
            "type": "string"
          },
          {
            "name": "isSupport",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "list of user",
            "in": "body",
            "type": "array",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "balance": {
                    "type": "number",
                    "example": 0
                  },
                  "role": {
                    "type": "string",
                    "example": "USER"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "example": "+911234567891"
                  },
                  "id": {
                    "type": "string",
                    "example": "userid2"
                  },
                  "uid": {
                    "type": "number",
                    "example": 62
                  },
                  "name": {
                    "type": "string",
                    "example": "Dohn Joe"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/v1/user/{id}": {
      "put": {
        "tags": [
          "User Control"
        ],
        "summary": "Initialize a user",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Initialize a user to the database with the users information.",
            "schema": {
              "required": [
                "name"
              ],
              "properties": {
                "name": {
                  "type": "string",
                  "description": "name of the user"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Details of the new initialized user",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "username"
                },
                "phoneNumber": {
                  "type": "string",
                  "example": "+919999999999"
                },
                "role": {
                  "type": "string",
                  "example": "ASTROLOGER"
                },
                "id": {
                  "type": "string",
                  "example": "userid"
                },
                "uid": {
                  "type": "number",
                  "example": 69
                },
                "balance": {
                  "type": "number",
                  "example": 0
                },
                "callStatus": {
                  "type": "string",
                  "example": "online"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "User Control"
        ],
        "summary": "Delete a user from database",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "User Control"
        ],
        "summary": "Retrieve a user",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Details of the user",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "username"
                },
                "phoneNumber": {
                  "type": "string",
                  "example": "+919999999999"
                },
                "role": {
                  "type": "string",
                  "example": "ASTROLOGER"
                },
                "id": {
                  "type": "string",
                  "example": "userid"
                },
                "uid": {
                  "type": "number",
                  "example": 69
                },
                "balance": {
                  "type": "number",
                  "example": 0
                },
                "callStatus": {
                  "type": "string",
                  "example": "online"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "User Control"
        ],
        "summary": "Update users call status",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Update users call status if ONLIVESTREAM / ONCALL / ONCHAT or ONLINE.",
            "schema": {
              "required": [
                "op",
                "path",
                "value"
              ],
              "properties": {
                "name": {
                  "op": "string",
                  "description": "operation -> REPLACE"
                },
                "path": {
                  "type": "string",
                  "description": "path of the value to change ie status"
                },
                "value": {
                  "type": "string",
                  "description": "the new status ie ONLINE"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Details of the user",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "username"
                },
                "phoneNumber": {
                  "type": "string",
                  "example": "+919999999999"
                },
                "role": {
                  "type": "string",
                  "example": "ASTROLOGER"
                },
                "id": {
                  "type": "string",
                  "example": "userid"
                },
                "uid": {
                  "type": "number",
                  "example": 69
                },
                "balance": {
                  "type": "number",
                  "example": 0
                },
                "callStatus": {
                  "type": "string",
                  "example": "online"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
    "/v1/user/{id}/session/user/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "role": {
                  "example": "any"
                },
                "expiry": {
                  "example": "any"
                },
                "tokentype": {
                  "example": "any"
                },
                "deviceId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/user/{id}/session/user/{id}/{deviceId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "deviceId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/v1/channel/{channelType}": {
      "post": {
        "tags": [
          "Channel"
        ],
        "summary": "Start a new channel with particular channelType",
        "description": "",
        "parameters": [
          {
            "name": "channelType",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Initialize channel with: \n   channelId = hostId\n - host   = details of host starting the channel\n - if any channel exists\n - delete the current channel and initialize a new",
            "schema": {
              "required": [
                "channelId",
                "host"
              ],
              "properties": {
                "channelId": {
                  "type": "string",
                  "description": "identifier to uniquely determine single livestream"
                },
                "host": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "identifier of the host"
                    },
                    "uid": {
                      "type": "number",
                      "description": "identifier of the host for agora"
                    },
                    "name": {
                      "type": "string",
                      "description": "name of the host"
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Details of the newly created channel.",
            "schema": {
              "type": "object",
              "properties": {
                "createTs": {
                  "type": "number",
                  "example": 1711790359043
                },
                "channelStatus": {
                  "type": "string",
                  "example": "ACTIVE"
                },
                "channelId": {
                  "type": "string",
                  "example": "01HPKHMV21AYACVMEBFNZ406Q3"
                },
                "host": {
                  "type": "object",
                  "properties": {
                    "uid": {
                      "type": "number",
                      "example": 63
                    },
                    "role": {
                      "type": "string",
                      "example": "USER"
                    },
                    "balance": {
                      "type": "number",
                      "example": 0
                    },
                    "name": {
                      "type": "string",
                      "example": "Vaibhav"
                    },
                    "phoneNumber": {
                      "type": "string",
                      "example": "+917722846537"
                    },
                    "id": {
                      "type": "string",
                      "example": "01HPKHMV21AYACVMEBFNZ406Q3"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
    "/v1/channel/{channelType}/{channelId}/event": {
      "post": {
        "tags": [
          "Channel"
        ],
        "summary": "Send an event to a channel/livestream joined -> Gift or Message",
        "description": "",
        "parameters": [
          {
            "name": "channelType",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "channelId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Send only message if sending a message or send giftId if sending a gift",
            "schema": {
              "required": [
                "userId"
              ],
              "properties": {
                "userId": {
                  "type": "string",
                  "description": "Identifier of the user sending the event"
                },
                "giftId": {
                  "type": "string",
                  "description": "identifier of the gift"
                },
                "message": {
                  "type": "string",
                  "description": "The message to send to the livestream"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response.",
            "schema": {
              "type": "object",
              "properties": {
                "event": {
                  "type": "string",
                  "example": "MessageOrGiftData"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/v1/channel/{channelType}/{channelId}/host/": {
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "channelType",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "channelId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "op": {
                  "example": "any"
                },
                "path": {
                  "example": "any"
                },
                "value": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/v1/support-case/user/{userId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "status",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "status",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/v1/session/user/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "role": {
                  "example": "any"
                },
                "expiry": {
                  "example": "any"
                },
                "tokentype": {
                  "example": "any"
                },
                "deviceId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/session/user/{id}/{deviceId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "deviceId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}